requirements.txt
torch
torchvision
xgboost
rasterio
numpy
flask
django
djangorestframework
requests
django-cors-headers